requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
PyJWT
google-cloud-storage